Simplify pass 2 (#2)
Instead of processing the generated ProcessedSourceLine objects in pass 2,
repeat pass 1 again, as other two pass assemblers does. This greatly simplifies
the code as pretty much of the code for pass 1 is reusable; performance is 
somewhat hurt but not dramatically (an expressions cache is used to help with that).

As a bonus, INCLUDEs in pass 2 only are now allowed
(even if they are still not very useful).